+++ plugin-hrm-form/src/components/case/Case.tsx
@@ -50,7 +50,7 @@ const getNameFromContact = contact => {
 
 const getNameFromForm = form => {
   const { firstName, lastName } = form.childInformation;
-  return formatName(`${firstName.value} ${lastName.value}`);
+  return formatName(`${firstName} ${lastName}`);
 };
 
 const Case: React.FC<Props> = props => {
