+++ plugin-hrm-form/src/components/case/helpers.js
@@ -1,0 +1,15 @@
+export function editNestedField(original, parents, name, change) {
+  if (parents.length === 0) {
+    return {
+      ...original,
+      [name]: {
+        ...original[name],
+        ...change,
+      },
+    };
+  }
+  return {
+    ...original,
+    [parents[0]]: editNestedField(original[parents[0]], parents.slice(1), name, change),
+  };
+}
